@@ -19,11 +19,11 @@
   },
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
-    "axios": "^1.3.1",
-    "dompurify": "^2.4.3",
-    "eslint": "^8.33.0",
+    "axios": "^1.3.5",
+    "dompurify": "^3.0.2",
+    "eslint": "^8.38.0",
     "lodash": "^4.17.21",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
@@ -32,16 +32,16 @@
     "react-dom": "^18.2.0",
     "react-easy-swipe": "0.0.23",
     "react-scripts": "5.0.1",
-    "react-svg-pan-zoom": "^3.12.0",
-    "sass": "^1.58.0"
+    "react-svg-pan-zoom": "^3.12.1",
+    "sass": "^1.62.0"
   },
   "devDependencies": {
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.6.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-jam3": "^0.2.3",
     "eslint-plugin-prettier": "^4.2.1",
-    "npm-check-updates": "^16.6.3",
-    "prettier": "^2.8.3",
+    "npm-check-updates": "^16.10.8",
+    "prettier": "^2.8.7",
     "react-app-rewired": "^2.2.1",
     "sass-lint": "^1.13.1",
     "sass-lint-webpack": "^1.0.4"
