@@ -26,7 +26,7 @@ function MetroZoneFinder() {
         {autoCompleteState.ticketMode ? (
           <div className="wmnds-m-b-lg wmnds-m-t-md">
             <a
-              href="https://find-a-ticket.wmnetwork.co.uk/"
+              href="https://find-a-ticket.tfwm.org.uk/"
               className={`wmnds-btn wmnds-btn--link ${s.backLink}`}
             >
               <Icon iconName="general-chevron-right" /> Back to ticket finder
