@@ -5,16 +5,19 @@ import CloseButton from './CloseButton/CloseButton';
 import s from './SelectedServiceHeader.module.scss';
 import useMapControls from '../../../MetroZoneFinder/Map/customHooks/useMapControls';
 
-const SelectedServiceHeader = ({ autoCompleteState, autoCompleteDispatch, queryId }) => {
+function SelectedServiceHeader({ autoCompleteState, autoCompleteDispatch, queryId }) {
   const selectedServiceRef = useRef(null);
   const { resetMapStation } = useMapControls();
 
   const selectedService = autoCompleteState.selectedStations[queryId];
 
+  const tramstop = selectedService.stationDisplayName;
+
   const handleClick = () => {
     resetMapStation(selectedService, autoCompleteState.selectedStations);
     autoCompleteDispatch();
   };
+
   return (
     <>
       {/* Close disruption box */}
@@ -22,18 +25,17 @@ const SelectedServiceHeader = ({ autoCompleteState, autoCompleteDispatch, queryI
         className={`wmnds-grid wmnds-grid--justify-between wmnds-m-t-xs wmnds-m-b-md ${s.selectedStationBox}`}
         ref={selectedServiceRef}
       >
-        <strong className={`wmnds-col-auto ${s.selectedSummary}`}>
-          {selectedService.routeName || selectedService.stationName}
-        </strong>
+        <strong className={`wmnds-col-auto ${s.selectedSummary}`}>{tramstop}</strong>
 
         <CloseButton onClick={handleClick} />
       </div>
     </>
   );
-};
+}
 
 // PropTypes
 SelectedServiceHeader.propTypes = {
+  // eslint-disable-next-line react/forbid-prop-types
   autoCompleteState: PropTypes.objectOf(PropTypes.any).isRequired,
   autoCompleteDispatch: PropTypes.func.isRequired,
   queryId: PropTypes.number.isRequired,
