@@ -13,8 +13,11 @@ import SelectedServiceHeader from '../SelectedServiceHeader/SelectedServiceHeade
 import useHandleAutoCompleteKeys from '../customHooks/useHandleAutoCompleteKeys';
 import useAutoCompleteAPI from '../customHooks/useAutoCompleteAPI';
 
-const TrainAutoComplete = ({ id, label, queryId }) => {
+function TrainAutoComplete({ id, label, queryId }) {
   const [autoCompleteState, autoCompleteDispatch] = useContext(AutoCompleteContext);
+
+  console.log('autoCompleteState');
+  console.log(autoCompleteState);
 
   const resultsList = useRef(null);
   const debounceInput = useRef(null);
@@ -30,11 +33,12 @@ const TrainAutoComplete = ({ id, label, queryId }) => {
 
   const resArr = [];
   filteredResults.forEach((item) => {
-    const i = resArr.findIndex((x) => x.stationName === item.stationName);
+    const i = resArr.findIndex((x) => x.stationDisplayName === item.stationDisplayName);
     if (i <= -1) {
       resArr.push({
         crsCode: item.crsCode,
         stationName: item.stationName,
+        stationDisplayName: item.stationDisplayName,
         metroZone: item.metroZone,
         metroZoneSecond: item.metroZoneSecond,
         parking: item.parking,
@@ -118,7 +122,7 @@ const TrainAutoComplete = ({ id, label, queryId }) => {
       )}
     </div>
   );
-};
+}
 
 // PropTypes
 TrainAutoComplete.propTypes = {
