@@ -5,9 +5,9 @@ import s from './Icon.module.scss';
 // Due to weird Protocol errors with external SVGs the svg use doesn't work well with production builds
 // So we Ajax the SVG in with a snippet at the bottom of public/index.html
 
-// Icons can be found at: https://designsystem.wmnetwork.co.uk/styles/icons/
+// Icons can be found at: https://designsystem.tfwm.org.uk/styles/icons/
 
-const Icon = ({ className, iconName, size, color, title }) => {
+function Icon({ className, iconName, size, color, title }) {
   return (
     <svg
       className={`${className} ${(color && s[color]) || ''}`}
@@ -18,7 +18,7 @@ const Icon = ({ className, iconName, size, color, title }) => {
       <use xlinkHref={`#wmnds-${iconName}`} href={`#wmnds-${iconName}`} />
     </svg>
   );
-};
+}
 
 Icon.propTypes = {
   iconName: PropTypes.string.isRequired,
