@@ -5,7 +5,7 @@ import dompurify from 'dompurify';
 
 const { sanitize } = dompurify;
 
-const Message = ({ type, title, message, showRetry, retryCallback }) => {
+function Message({ type, title, message, showRetry, retryCallback }) {
   let iconName;
   switch (type) {
     case 'error':
@@ -36,7 +36,7 @@ const Message = ({ type, title, message, showRetry, retryCallback }) => {
       </div>
     </div>
   );
-};
+}
 
 Message.propTypes = {
   type: PropTypes.string,
